[onert/ruy] Move operation visitors to separate layer files
This commit moves Conv2D and FullyConnected visitor implementations from KernelGenerator.cc to their respective operation layer files.
It also updates Validator.h to use visitor declarations instead of inline implementations.

ONE-DCO-1.0-Signed-off-by: Hyeongseok Oh <hseok82.oh@samsung.com>
